+++ ChangeLog.md
@@ -7,6 +7,7 @@
     Release date: 
 
     - Add shown parameter to the tab definition, thus bumping minor candidate version number
+    - Fix tab-pane margin top to not be added to tabbed-navs margin bottom
 
 ### 1.4.0