+++ ChangeLog.md
@@ -9,6 +9,7 @@
     - Add shown parameter to the tab definition, thus bumping minor candidate version number
     - Fix tab-pane margin top to not be added to tabbed-navs margin bottom
     - package.js refactoring
+    - Fix next tab activating computing
 
 ### 1.4.0